Change basic requirements for testnet
Enable Proof Of Work on the testnet, this is left enabled so we can move the chain if
for some reason it gets stuck. Proof Of Stake relies on the existence of a relatively
high number of staking transactions so that it can stake enough to confirm previously
staked blocks. In the event that staking fails to do this we will use Proof Of Work
to confirm the staked blocks and allow the coins to be staked again.