qt: use series of pngs for spinner
Use a series of .png frames for the spinner instead of a .mng. `mng` is
an obscure image format and is not built by default into Qt5 (future
planning).

This appears to improve the crispness of the spinner as well.

Does not noticeably increase the size (still ~27k) and the code is not
more complicated either.