Implement scrape addresses
Enable setting a scrape address for any in wallet address:
This will scrape all stake rewards at a transaction level and send the reward to
the scrape address.
Scrape addresses are stored in their own database and can be set, listed or
deleted using RPC via daemon or the in QT RPC console.