Partial scrape address control functionality for QT wallet.
full/proper implementation would require a fairly significant refactoring of the
addressbookpage.cpp/.h addresstablemodel.cpp/.h files so to avoid that we hide
the scrape address field on the sending tab of the address book ("Address Book"),
only displaying it on the receive tab ("Receive coins").

To avoid refactoring the afforementioned files as well as editaddressdialog.cpp/.h
files and accompanied forms the edit address option does not allow you to edit
the scrape address but double clicking on the field works as expected.