Don't stop users from using 0.0.0.0 as the a bind address
While 0.0.0.0 is already the default bind address for IPv4 blocking
it breaks old conf files that are configured using this.

Instead of removing the IsValid check on bind addresses don't check
if the address is a 0 address during init (allow other IsValid checks
to function as they always have).

0, 0.0 and 0.0.0 addresses are all expanded to 0.0.0.0 (confirmed by
printing to string during tests) while anything longer is obviously
not an IPv4 address and should be checked like normal.